refactor: #95 - Melhora nomemclatura e organização dos arquivos da atualização cadastral
